+++ src/bract/ring/util.clj
@@ -55,3 +55,10 @@ Request: %s"
   "Do nothing at all, returning nil."
   [& args]
   nil)
+
+
+(defn elapsed-millis
+  "Return elapsed milliseconds (with decimal places) since start time in nanoseconds."
+  ^double [^long start-nanos]
+  (let [total-ns (unchecked-subtract (System/nanoTime) start-nanos)]
+    (double (/ total-ns 1e6))))